added instruct pix2pix example
